Move cookie reading into identity component
