+++ src/components/Identity/createGetEcidFromCookie.js
