+++ src/components/Identity/index.js
@@ -37,6 +37,7 @@ import createIdentityRequest from "./getIdentity/createIdentityRequest.js";
 import createIdentityRequestPayload from "./getIdentity/createIdentityRequestPayload.js";
 import injectAppendIdentityToUrl from "./appendIdentityToUrl/injectAppendIdentityToUrl.js";
 import createGetIdentityOptionsValidator from "./getIdentity/createGetIdentityOptionsValidator.js";
+import createGetEcidFromCookie from "./createGetEcidFromCookie.js";
 
 const createIdentity = ({
   config,
@@ -123,6 +124,11 @@ const createIdentity = ({
     thirdPartyCookiesEnabled,
     areThirdPartyCookiesSupportedByDefault,
   });
+  const getEcidFromCookie = createGetEcidFromCookie({
+    orgId,
+    cookieJar: loggingCookieJar,
+    logger,
+  });
   return createComponent({
     addEcidQueryToPayload,
     addQueryStringIdentityToPayload,
@@ -136,6 +142,7 @@ const createIdentity = ({
     logger,
     config,
     getIdentityOptionsValidator,
+    getEcidFromCookie,
   });
 };
