+++ src/core/componentCreators.js
@@ -17,7 +17,6 @@ governing permissions and limitations under the License.
 export { default as activityCollector } from "../components/ActivityCollector/index.js";
 export { default as audiences } from "../components/Audiences/index.js";
 export { default as context } from "../components/Context/index.js";
-export { default as cookieReader } from "../components/CookieReader/index.js";
 export { default as rulesEngine } from "../components/RulesEngine/index.js";
 export { default as eventMerge } from "../components/EventMerge/index.js";
 export { default as mediaAnalyticsBridge } from "../components/MediaAnalyticsBridge/index.js";
